发现一个问题,因为有自动封包和解包,会导致生成一些"短命"对象 更改了一个变量的类型: long --> int

+++ trunk/XRace_V2/src/com/sa/xrace/client/scene/GLWorld.java
@@ -66,8 +66,8 @@ public class GLWorld {
         collisionMap.prepare();
     }
 
-    public void draw(GL10 gl, long timeElapsed) {
-    	final int time = (int)timeElapsed;
+    public void draw(GL10 gl, int timeElapsed) {
+    	final int time = timeElapsed;
         myCar = ObjectPool.myCar;
         if (StateValuePool.isBeginWait) {
             myCar.updateSpeedByKeyboard(time / 80);
